requirements.txt
mediapipe
opencv-python
pillow